+++ languages/python/component-template/tests/test_main.py
@@ -1,0 +1,4 @@
+import @mainPackage@
+
+def test_main() -> None:
+    @mainPackage@.main()